Domotic/api/requirements.txt
fastapi
uvicorn
paho-mqtt
dotenv